move app out of src
