+++ App.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { SafeAreaView, StyleSheet, View } from "react-native";
-import Playground from "./components/Playground";
+import Playground from "./src/components/Playground";
 
 export default function App() {
   return (
